fix(heat): correct rendering direction and partial-level display
- changed gradient orientation from horizontal to vertical for heat bar
- allowed float-based heat levels for smoother transitions
- improved badge rendering with fractional heat support (half-levels)
- added paragraph numbering tooltips and better DOM key tracking
- fixed missing paragraph element handling with clearer warnings

+++ pkg/schema/entities.go
@@ -4,10 +4,10 @@ type Summary struct {
 	Characters []Character `json:"characters" jsonschema_description:"List of extracted characters with their attributes and actions"`
 	Timeline   []Timeline  `json:"timeline" jsonschema_description:"Chronological sequence of dated events extracted from the story"`
 
-	Chapters map[string]bool `json:"chapters" jsonschema_description:"List of chapter IDs visited (filled by backend)"`
-	Heat     map[string]int  `json:"heat,omitempty" jsonschema_description:"Mapping of paragraph (starting from 1) with a level of 0-3 based on sexual activity intensity"`
+	Chapters map[string]bool    `json:"chapters" jsonschema_description:"List of chapter IDs visited (filled by backend)"`
+	Heat     map[string]float64 `json:"heat,omitempty" jsonschema_description:"Mapping of paragraph (starting from 1) with a level of 0-3 based on sexual activity intensity"`
 
-	StoredHeat map[string]map[string]int `json:"stored_heat,omitempty" jsonschema_description:"Backend storage of heat maps per chapter ID"`
+	StoredHeat map[string]map[string]float64 `json:"stored_heat,omitempty" jsonschema_description:"Backend storage of heat maps per chapter ID"`
 }
 
 type Character struct {
